parallelize json files too

+++ step4_makejson.py
@@ -3,14 +3,17 @@ from helperstuff.templates import TemplatesFile, templatesfiles
 import json
 
 def makejson(*args):
-    templatesfile = TemplatesFile(*args)
-    if templatesfile.copyfromothertemplatesfile is not None: return
+  templatesfile = TemplatesFile(*args)
+  if templatesfile.copyfromothertemplatesfile is not None: return
+  with KeepWhileOpenFile(templatesfile.jsonfile()+".tmp", message=LSB_JOBID()) as kwof:
     print templatesfile
+    if not kwof: return
+    if os.path.exists(templatesfile.jsonfile()): return
 
     jsonstring = json.dumps(templatesfile.getjson(), sort_keys=True, indent=4, separators=(',', ': '))
     filename = templatesfile.jsonfile()
     with open(filename, "w") as f:
-        f.write(jsonstring)
+      f.write(jsonstring)
 
 if __name__ == "__main__":
     for templatesfile in templatesfiles:
